requirements.txt
asgiref==3.2.7
boto3==1.14.20
botocore==1.17.20
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.0.7
django-appconf==1.0.4
django-cors-headers==3.5.0
django-environ==0.4.5
django-heroku==0.3.1
django-imagekit==4.0.2
django-s3-storage==0.13.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
pilkit==2.0
Pillow==7.2.0
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.2.0
